+++ BlibliMovies2/src/Controller/FilmGenre/FilmGenreDelete.java
@@ -45,7 +45,11 @@ public class FilmGenreDelete extends HttpServlet {
         try{
             FilmGenre filmGenre = filmService.getFilmGenre(request.getParameter("id"), (int)request.getSession().getAttribute("storeid"));
 
-            filmService.deleteFilmGenre(filmGenre.getId() + "", (int)request.getSession().getAttribute("storeid"));
+            if(filmGenre.isStatus()){
+                filmService.deleteFilmGenre(filmGenre.getId() + "", (int)request.getSession().getAttribute("storeid"));
+            } else{
+                filmService.retrieveFilmGenre(filmGenre.getId() + "", (int)request.getSession().getAttribute("storeid"));
+            }
 
             address = "/view/database/success.jsp";
             request.setAttribute("title", "Film Genre");
