+++ BlibliMovies2/src/Controller/FnB/FnBDelete.java
@@ -43,7 +43,11 @@ public class FnBDelete extends HttpServlet {
         try{
             FnB fnb = fnbService.getFnB(request.getParameter("id"), (int)request.getSession().getAttribute("storeid"));
 
-            fnbService.deleteFnB(fnb.getId() + "", (int)request.getSession().getAttribute("storeid"));
+            if(fnb.isStatus()){
+                fnbService.deleteFnB(fnb.getId() + "", (int)request.getSession().getAttribute("storeid"));
+            } else {
+                fnbService.retrieveFnB(fnb.getId() + "", (int)request.getSession().getAttribute("storeid"));
+            }
 
             address = "/view/database/success.jsp";
             request.setAttribute("title", "FnB");
