+++ BlibliMovies2/src/Controller/FnBSize/FnBSizeDelete.java
@@ -45,7 +45,11 @@ public class FnBSizeDelete extends HttpServlet {
         try{
             FnBSize fnbSize = fnbService.getFnBSize(request.getParameter("id"), (int)request.getSession().getAttribute("storeid"));
 
-            fnbService.deleteFnBSize(fnbSize.getId() + "", (int)request.getSession().getAttribute("storeid"));
+            if(fnbSize.isStatus()){
+                fnbService.deleteFnBSize(fnbSize.getId() + "", (int)request.getSession().getAttribute("storeid"));
+            } else {
+                fnbService.retrieveFnBSize(fnbSize.getId() + "", (int)request.getSession().getAttribute("storeid"));
+            }
 
             address = "/view/database/success.jsp";
             request.setAttribute("title", "FnB Size");
