+++ BlibliMovies2/src/Controller/ScreeningTime/ScreeningTimeDelete.java
@@ -37,7 +37,13 @@ public class ScreeningTimeDelete extends HttpServlet{
         }
 
         try{
-            filmService.deleteScreeningTime(request.getParameter("id"), request.getParameter("filmid"), (int)request.getSession().getAttribute("storeid"));
+            ScreeningTime screeningTime = filmService.getScreeningTime(request.getParameter("id"), request.getParameter("filmid"), (int)request.getSession().getAttribute("storeid"));
+
+            if(screeningTime.isStatus()){
+                filmService.deleteScreeningTime(request.getParameter("id"), request.getParameter("filmid"), (int)request.getSession().getAttribute("storeid"));
+            } else {
+                filmService.retrieveScreeningTime(request.getParameter("id"), request.getParameter("filmid"), (int)request.getSession().getAttribute("storeid"));
+            }
 
             String address = "/view/database/success.jsp";
             request.setAttribute("title", "Screening Time");
